Reuse visitArticleNoOkta logic in benchmarking
